% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chas.R
\name{chas_make_dictionary}
\alias{chas_make_dictionary}
\title{Create a data dictionary of all variables}
\usage{
chas_make_dictionary(chas_dictionary_file)
}
\arguments{
\item{chas_dictionary_file}{An excel file in each CHAS unzip folder with a sheet defining geographies all variables in all tables.}
}
\value{
a dataset containing a list of all variables in all tables.
}
\description{
Create a data dictionary of all variables
}
