% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chas.R
\name{get_chas}
\alias{get_chas}
\title{Get HUD CHAS data for a geography.}
\usage{
get_chas(geography, year = NULL, state = NULL, county = NULL)
}
\arguments{
\item{geography}{The geography of your data:
\itemize{
\item \code{"state"}
\item \code{"county"}
\item \code{"tract"}: census tracts
\item \code{"place"}: census places (cannot narrow by county)
\item \code{"counties by place"}: counties divided into places
\item \code{"mcd"}: minor civil divisions (such as county districts)
\item \code{"consolidated cities"}: consolidated city-county gov't
}}

\item{year}{Numeric end year of the data (defaults to 2020)}

\item{state}{Character string of two-digit state FIPS code (ex: "47")}

\item{county}{Character string of three-digit county FIPS code (ex: "157")}
}
\value{
A tibble of the requested data.
}
\description{
Get HUD CHAS data for a geography.
}
\examples{
get_chas("state", state = "47") |> head()
}
