% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychas-package.R
\docType{package}
\name{tidychas-package}
\alias{tidychas}
\alias{tidychas-package}
\title{tidychas: Access Full HUD CHAS Data with R}
\description{
tidychas is an R package for accessing and tidying HUD's CHAS data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sj-io/tidychas}
  \item Report bugs at \url{https://github.com/sj-io/tidychas/issues}
}

}
\author{
\strong{Maintainer}: Sarah Johnson \email{sarahviamemphis@gmail.com}

}
\keyword{internal}
